refactor(optionsLoader): Fix
