+++ src/context.ts
@@ -1,36 +1,12 @@
 import { PrismaClient } from "@prisma/client";
 import { ApolloServerExpressConfig } from "apollo-server-express";
-import DataLoader from "dataloader";
 import { JsonPlaceholderClient } from "./api/JsonPlaceholderClient";
+import { optionsLoaderFactory } from "./graphql/loaders";
 
 export type Context = {
   jsonPlaceholderClient: JsonPlaceholderClient,
   prismaClient: PrismaClient,
-  optionsLoader: ReturnType<typeof optionsLoader>,
-};
-
-const optionsLoader = (prisma: PrismaClient) => {
-  return new DataLoader(async (ids: number[]) => {
-    const res = await prisma.option.findMany({
-      where: {
-        questionId: {
-          in: ids,
-        },
-      },
-    });
-    const optionsByQuestionId = new Map();
-    res.forEach((option) => {
-      const questionId = option.questionId;
-      if (optionsByQuestionId.has(questionId)) {
-        const current = optionsByQuestionId.get(questionId);
-        optionsByQuestionId.set(questionId, [...current, option]);
-        return;
-      }
-      optionsByQuestionId.set(questionId, [option]);
-    });
-    const options = [...optionsByQuestionId.values()];
-    return options;
-  });
+  optionsLoader: ReturnType<typeof optionsLoaderFactory>,
 };
 
 export const createContext: ApolloServerExpressConfig["context"] = () => {
@@ -46,6 +22,6 @@ export const createContext: ApolloServerExpressConfig["context"] = () => {
   return {
     jsonPlaceholderClient: new JsonPlaceholderClient(),
     prismaClient,
-    optionsLoader: optionsLoader(prismaClient),
+    optionsLoader: optionsLoaderFactory(prismaClient),
   };
 };
